fix Int<=>Double problem

+++ src/main/kotlin/jp/trap/conqest/listeners/ListenerDistrictCore.kt
@@ -28,6 +28,6 @@ class ListenerDistrictCore(private val gameManager: GameManager) : Listener {
         val game = gameManager.getGame(player) ?: return
         val districts = game.field.districts
         val district = districts.firstOrNull { district -> district.coreLocation.block == event.block.location.block }
-        district?.changeHP(100 - event.progress.toDouble() * 100)
+        district?.changeHP((100 - event.progress.toDouble() * 100).toInt())
     }
 }
